+++ tests/unit_test/test_command_server.py
@@ -6,9 +6,12 @@ import socket
 import tempfile
 import threading
 import time
+from typing import Any
 
 from PIL import Image
 
+import actfw_core
+from actfw_core.application import Application
 from actfw_core.command_server import CommandServer
 from actfw_core.schema.agent_app_protocol import CommandKind, CommandRequest, CommandResponse, RequestId, Status
 
@@ -122,3 +125,30 @@ def test_socket_timeout_keeps_command_server_running() -> None:
         cmd.stop()
         cmd.join()
         tmpdir.cleanup()
+
+
+def test_command_server_terminate() -> None:
+    # try to catch assert error
+    exception = None
+
+    def excepthook(args: Any) -> None:
+        nonlocal exception
+        exception = args.exc_value
+
+    threading.excepthook = excepthook
+
+    # Actcast application
+    app = actfw_core.Application()
+
+    # CommandServer (for `Take Photo` command)
+    cmd = actfw_core.CommandServer("/tmp/test.sock")
+    app.register_task(cmd)
+
+    # Start application
+    th = threading.Thread(target=lambda: app.run())
+    th.start()
+    # terminate CommandServer task without calling `update_image`
+    time.sleep(0.1)
+    app.stop()
+    th.join()
+    assert exception is None
